+++ CHANGELOG.md
@@ -10,6 +10,7 @@ develop
 - Improve text finder:
     - Now, “Find All” action also highlights all matched strings in the editor, and thereby “Highlight” action is removed.
     - On Yosemite and later, a visual feedback is shown when the seach wrapped.
+    - Keep selected range after “Replace All” with in-selection option.
     - Revert “Use selection for Replace” action to allow using an empty string.
     - Tweak layout and style.